pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.6.0",
    "coverage>=7.13.0",
    "django>=5.2.4",
    "django-browser-reload>=1.21.0",
    "django-countries>=8.2.0",
    "django-tailwind>=4.2.0",
    "flake8>=7.3.0",
    "flake8-isort>=6.1.2",
    "onlinepayments-sdk-python3>=6.4.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
]
